+++ crates/web-client/src/models/mod.rs
@@ -87,7 +87,6 @@ pub mod provers;
 pub mod public_key;
 pub mod rpo256;
 pub mod script_builder;
-pub mod secret_key;
 pub mod signature;
 pub mod signing_inputs;
 pub mod storage_map;